Added mouse based color isolation script, fixed picture test
Supplying the full path to a resource shouldn't be necessary, and will
break code when we try to share, so if possible leave it as just a
relative path.

The vision stuff took me a bit to get right, I had the X and Y image
axes flipped in my head, causing my masks to do strange things.